wip: postgresql settings for prod
