+++ conf/application.conf
@@ -26,7 +26,7 @@ jvm.opts = [
 
 // DB configuration
 db.default.driver=org.h2.Driver
-db.default.url="jdbc:h2:mem:minesweeper-api;MODE=MYSQL;INIT=RUNSCRIPT FROM 'classpath:scripts/init-db-h2.sql'"
+db.default.url="jdbc:h2:mem:minesweeper-api;MODE=PostgreSQL;INIT=RUNSCRIPT FROM 'classpath:scripts/init-db.sql'"
 
 scalikejdbc.global.loggingSQLAndTime {
   enabled = true
