+++ conf/application.prod.conf
@@ -1,1 +1,8 @@
 include "application.conf"
+
+play.evolutions.db.default.autoApply=true
+
+db.default {
+  driver = org.postgresql.Driver
+  url = ${DATABASE_URL}
+}
